Reverting all changes to base code
